mod: Add to list of exported functions.

+++ kernel/inc/string.h
@@ -1,7 +1,7 @@
 #ifndef STRING_H
 #define STRING_H
 
-#include <types.h>
+#include <stdint.h>
 
 /**
  * Calculates the length of a string.
@@ -9,7 +9,7 @@
  * @param str the string to calculate the length of
  * @return the length of the string
  */
-uint32_t strlen(const char *str);
+size_t strlen(const char *str);
 
 /**
  * Calculates the length of a wide string.
@@ -17,7 +17,7 @@ uint32_t strlen(const char *str);
  * @param str the wide string to calculate the length of
  * @return the length of the wide string
  */
-uint32_t wcslen(const short *str);
+size_t wcslen(const short *str);
 
 /**
  * Checks to see if two strings are identical
@@ -27,7 +27,7 @@ uint32_t wcslen(const short *str);
  */
 int strcmp(const char *str1, const char *str2);
 
-int strncmp(const char *str1, const char *str2, uint32_t num);
+int strncmp(const char *str1, const char *str2, size_t num);
 
 char *strchr(const char *s, int c);
 
@@ -35,10 +35,10 @@ char *strcpy(char *dest, const char *src);
 
 char *strncpy(char *dest, const char *src, size_t n);
 
-void *memcpy(void *dest, const void *src, uint32_t n);
+void *memcpy(void *dest, const void *src, size_t n);
 
-void *memset(void *s, int c, uint32_t n);
+void *memset(void *s, int c, size_t n);
 
-void *memmove(void *dst, const void *src, uint32_t n);
+void *memmove(void *dst, const void *src, size_t n);
 
 #endif
